First prep. to write all measures to file
So far implemented it for one observable file. Have to rewrite this so
that we write out measurements after a configuration has been processed.
At the moment we write the measurements at the end of the program. In
case of a crash that would mean, that we lose all progress.