Modified meas writeout so that it happens after every meas
The writeout of single measures not gets prepared before the run starts
(overwrite the file if its there and insert the description at the
beginning) and then after every measurement (every configuration) write
out the observables into that (one) file.
This does not happen for box counting or for all radii, yet. This will
need a little bit of rewrite and additional functions for these two
observables.