Writeout stdout results without Jackknife if nmeas<2
To determine the f(a) error using Jackknife we have to determine the
Radius of the largest cluster for only one configuration at a time. As
we are using the stdout results printed from findcluster in the Python
script, we had to modify it, so that it writes out the configuration
results instead of Jackknife (which gives NaNs) when nmeas=1.