Cleanup of radius calculation
There are now two functions, one to calculate the radius of one cluster
and one to calculate it for all clusters using the former function. The
distrinction between non-percolating and percolating cluster
calculations is now just calling it with the respective clusterid. This
is implemented in findcluster_obs.cpp now.